requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-whois>=0.8.0
urllib3>=1.26.0